% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{barplot}
\alias{barplot}
\alias{barplot.EnrichmentFET}
\title{barplot for fet enrichment test}
\usage{
barplot(x, ...)

\method{barplot}{EnrichmentFET}(
  x,
  num = 6,
  greater = 3,
  less = 3,
  text.size = 10,
  fill = "gray20",
  panel.ratio = 1.2,
  title,
  ...
)
}
\arguments{
\item{x}{EnrichmentFET. Returned by \code{\link{enrichment_test}}
with test equals to fet.}

\item{...}{other params not supported}

\item{num}{integer. Number of categories to display. Use it when alternative
equals to greater, less, or two.sided.}

\item{greater}{integer. Number of categories which are greater. Use it when
alternative equals to both.}

\item{less}{integer. Number of categories which are less. Use it when
alternative equals to both.}

\item{text.size}{integer}

\item{fill}{character. Bar fill color}

\item{panel.ratio}{numeric. The ratio of left and right panels.}

\item{title}{character. The title of the plot.}
}
\value{
a ggplot object
}
\description{
This function makes bar plot for top most positively and negatively
enriched levels.
}
\examples{
lpd = transform_by_sample(lipidome, function(x) log(x/sum(x)))
design = model.matrix(~Treatment * Timepoint + Subject, data = lpd$pdata)
fit = model_fit(lpd, design, "TreatmentMed:TimepointPre", "limma")
en = enrichment_test(lpd, fit, "class", "fet")
barplot(en)
en = enrichment_test(lpd, fit, "class", "fet", alternative = "greater")
barplot(en)
}
\seealso{
\code{\link{enrichment_test}}
}
\author{
Chenghao Zhu
}
