% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSet-methods.R
\name{dim,HTSet-method}
\alias{dim,HTSet-method}
\title{Get dimemsions}
\usage{
\S4method{dim}{HTSet}(x)
}
\arguments{
\item{x}{A \code{\link{HTSet-class}} or derived class object.}
}
\value{
A integer vector with length of 2. The first is the number of
features, and the second is number of samples.
}
\description{
Get the dimensions of the edata of an HTSet object
}
\seealso{
\code{\link{HTSet-class}}
}
