% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{ecdf}
\alias{ecdf}
\alias{ecdf.EnrichmentKST}
\title{ecdf plot for kst enrichment test}
\usage{
ecdf(x, ...)

\method{ecdf}{EnrichmentKST}(
  x,
  level,
  alternative = c("greater", "less"),
  colors = c("blue", "red"),
  ...
)
}
\arguments{
\item{x}{EnrichmentKST. Returned by \code{\link{enrichment_test}}
when test equals to kst.}

\item{...}{other arguments not supported}

\item{level}{character. The level to plot the ecdf plot. Must be in the
group.}

\item{alternative}{character. Either greater or less}

\item{colors}{character}
}
\value{
a ggplot2 object
}
\description{
makes a ecdf plot for enrichment test result returned by
\code{\link{enrichment_test}} with test equals to kst.
}
\examples{
lpd = transform_by_sample(lipidome, function(x) log(x/sum(x)))
design = model.matrix(~Treatment * Timepoint + Subject, data = lpd$pdata)
fit = model_fit(lpd, design, "TreatmentMed:TimepointPre", "limma")
en = enrichment_test(lpd, fit, "class", "kst")
ecdf(en, level = "PC", alternative = "greater")

en = enrichment_test(lpd, fit, "class", "kst", alternative = "greater")
ecdf(en, level = "PC")

en = enrichment_test(lpd, fit, "class", "kst", alternative = "two.sided")
ecdf(en, level = "PC")

}
\seealso{
\code{\link{enrichment_test}}
}
\author{
Chenghao Zhu
}
