% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment_test}
\alias{enrichment_test}
\title{Enrichment test}
\usage{
enrichment_test(
  object,
  fit,
  group,
  test = c("fet", "kst"),
  alternative = c("both", "greater", "less", "two.sided"),
  p.cutoff
)
}
\arguments{
\item{object}{\code{\link{HTSet-class}}}

\item{fit}{ModelFit. The result of \code{\link{model_fit}}}

\item{group}{character. The fdata column of object that will be used as the
categories to perform enrichment analysis.}

\item{test}{character. Must be "fet" for Fisher's exact test or "kst" for
Kolmogorov-Smirnov test.}

\item{alternative}{character. One of both, greater, less, or two.sided.
Default is both, that both greater and less were calculated and returned.}

\item{p.cutoff}{numeric. The p-value cutoff to define increase or decreasing.
Only useful with fet. The default value of this argument is 0.05 when
alternative equals to two.sided, and is 1 otherwise.}
}
\value{
A list-like S3 object. \code{pval} is the pvalues for enrichment of
each category. For fet, the \code{odds.ratio} has the odds ratio for each
category, and the \code{matrix} has the N, m, n, k, and x values. For kst,
\code{adj.p.val} has the transformed p-values under each of the alternative
hypothesis of greater or less.
}
\description{
Perform enrichment test of genes/features in categories with
Fisher's exact test or Kolmogorov-Smirnov Tests.
}
\details{
For Fisher's exact test, the question being answered is, whether the number
of genes/features in category A that are increased (or decreased) in
treatment different from the background. In other words, the Fisher's exact
test is used to test the independency of gene/feature category A and whether
genes/features are increased (or decreased).

For Kilmogorov-Smirnov test, a similar hypothesis is tested. The p-values
are first tranformed into one-tail. The transformation is done using the
statistic values returned by \code{\link{model_fit}}. The transformed
p-values are then compared to an arithmetic sequence between 0 and 1 using
\code{\link{ks.test}}.
}
\examples{
design = model.matrix(~ Condition, data = exrna$pdata)
coef = "ConditionSystemic Lupus Erythematosus"
fit = model_fit(exrna, design, coef, engine = "limma", args = list(voom = TRUE))
en = enrichment_test(exrna, fit, "gene_type", "fet")
en = enrichment_test(exrna, fit, "gene_type", "kst")

}
\seealso{
\code{\link{HTSet-class}}
\code{\link{model_fit}}
\code{\link{fisher.test}}
\code{\link{phyper}}
\code{\link{ks.test}}
}
\author{
Chenghao Zhu
}
