% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exrna}
\alias{exrna}
\title{exRNA from human isolated HDL}
\format{A \code{\link{HTSet-class}} object}
\usage{
data(exrna)
}
\description{
extra-cellular small RNA-seq gene counts from isolated HDL
from 3 human patients with systemic lupus erythematosus and 3 healthy
control.
}
\references{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE121865
}
\keyword{datasets}
