% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSet-accessors.R
\name{$,HTSet-method}
\alias{$,HTSet-method}
\alias{$}
\alias{$<-}
\alias{$<-,HTSet-method}
\alias{[[<-}
\alias{[[<-,HTSet-method}
\alias{[,HTSet-method}
\alias{[}
\title{Extract or replace parts of an HTSet object}
\usage{
\S4method{$}{HTSet}(x, name)

\S4method{$}{HTSet}(x, name)

\S4method{$}{HTSet}(x, name) <- value

\S4method{[[}{HTSet}(x, i, j, ...) <- value

\S4method{[}{HTSet}(x, i, j, k, l, ..., drop = FALSE)
}
\arguments{
\item{x}{\code{\link{HTSet-class}}.}

\item{name}{The name of the slot.}

\item{value}{A replacement value for this slot, which must match the slot
definition.}

\item{i, j}{indices for features (i) and samples (j) to extract.}
}
\description{
Operators "$" and [[ extract or replace a slot, while [ extracts
a subset of the \code{\link{HTSet-class}} object in a similar manner of
\code{\link{subset_features}} and \code{\link{subset_samples}}.
}
\seealso{
\code{\code{HTSet-class}}
}
