% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSet-methods.R
\name{featureNames,HTSet-method}
\alias{featureNames,HTSet-method}
\alias{featureNames}
\alias{featureNames<-,HTSet-method}
\alias{featureNames<-}
\title{Get or set the feature names of an HTSet object}
\usage{
\S4method{featureNames}{HTSet}(x)

\S4method{featureNames}{HTSet}(x) <- value
}
\arguments{
\item{x}{\code{\link{HTSet-class}}}

\item{value}{A character vector. The length must match number of features.}
}
\value{
A character vector.
}
\description{
Get or set the feature names of an \code{\link{HTSet-class}}
object.
}
