% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome-summarize.R
\name{format_lipid_name}
\alias{format_lipid_name}
\title{format lipid names}
\usage{
format_lipid_name(x)
}
\arguments{
\item{x}{character.}
}
\value{
character
}
\description{
Format lipid names into a unified format, such as PC 36:1. This
is useful for downstream processing such as calculation $EOD_18$ or ACL.
}
\author{
Chenghao Zhu
}
