% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{hist.ModelFit}
\alias{hist.ModelFit}
\title{p value histogram}
\usage{
\method{hist}{ModelFit}(
  x,
  vline,
  bins = 40,
  color = "white",
  fill = "gray10",
  vline_linetype = "dashed",
  vline_size = 0.5,
  vline_color = "gray30",
  ...
)
}
\arguments{
\item{x}{ModelFit. Must be returned by \code{\link{model_fit}}}

\item{vline}{p value where vertical line to draw}

\item{bins}{numeric. Total number of bins}

\item{color}{character for bin outline colors}

\item{fill}{character for bin fill colors}

\item{vline_linetype}{character for vline linetype}

\item{vline_size}{numeric for vline size}

\item{vline_color}{character for vline color}

\item{...}{other args}
}
\value{
ggplot
}
\description{
create a histogram for p values of model fit result returned by
\code{\link{model_fit}}
}
\author{
Chenghao Zhu
}
