% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSet-class.R
\docType{class}
\name{HTSet-class}
\alias{HTSet-class}
\title{HTSet}
\description{
S4 class for storing high through-put experiment data. It has
three enssential components, the edata, fdata, and pdata. The edata is the
expression/abundance matrix of each feature in each sample. The features are
in rows and samples are in columns. The fdata is the feature data which
holds the information for each feature, such as gene names, protein IDs, or
metabolite names. The pdata is the phenotype data that describe additional
information for each sample, such as sex, condition, or treatment. The rows
in the edata must match the rows in the fdata, and the columns in the edata
must match the rows in the pdata.
}
\section{Slots}{

\describe{
\item{\code{edata}}{matrix. Must be numeric (integer or float)}

\item{\code{fdata}}{data.frame. The number of rows and row names must match those
of the edata.}

\item{\code{pdata}}{data.frame. The number of rows and row names must match the
number of columns and column names of edata.}

\item{\code{assay}}{list. Additional assay or study information.}
}}

\author{
Chenghao Zhu
}
