% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSet-class.R
\name{HTSet}
\alias{HTSet}
\title{HTSet S4 class constructur}
\usage{
HTSet(edata, fdata = NULL, pdata = NULL, assay = NULL)
}
\arguments{
\item{edata}{matrix}

\item{fdata}{data.frame, default is NULL}

\item{pdata}{data.frame, default is NULL}

\item{assay}{list,t default is NULL}
}
\value{
\code{\link{HTSet-class}}
}
\description{
The constructore for the \code{\link{HTSet-class}}.
}
\seealso{
\code{\link{HTSet-class}}
}
\author{
Chenghao Zhu
}
