% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipidome}
\alias{lipidome}
\title{lipidomics dataset from human isolated HDL}
\format{A \code{\link{HTSet-class}} object}
\usage{
data(lipidome)
}
\description{
A \code{\link{HTSet-class}} dataset of lipidomics analysis
from isolated HDL from 10 human subjects with 2x2 cross-over study design.
The data was obtained from the article published on \emph{Metabolomics} in
2019 (doi.org/10.1007/s11306-019-1579-1)
}
\references{
Zhu, C. et al., The HDL lipidome is widely remodeled by fast food versus
Mediterranean diet in 4 days. \emph{Metabolomics}, 2019 Aug 17;15(8):114.
doi: 10.1007/s11306-019-1579-1.
}
\keyword{datasets}
