% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome-summarize.R
\name{mz2molwt}
\alias{mz2molwt}
\title{m/z to molecular weight}
\usage{
mz2molwt(species, mz)
}
\arguments{
\item{species}{character value of the adduct ion species}

\item{mz}{numeric value of the m/z}
}
\value{
numeric
}
\description{
Calculate the molecular weight from m/z and adduct ion species.
The calculation is based on the adduct table extracted from the West Coast
Metabolomics Center's websit. See \code{\link{wcmc_adduct}}.
}
\seealso{
\code{\link{wcmc_adduct}}
}
\author{
Chenghao Zhu
}
