% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome-summarize.R
\name{nCarbons}
\alias{nCarbons}
\title{Extract the carbon chain length}
\usage{
nCarbons(x)
}
\arguments{
\item{x}{character. The lipid annotation name. Must be first cleaned by
\code{\link{format_lipid_name}}.}
}
\value{
integer
}
\description{
Extract the total number of carbons from a lipid annotation
name. For example, the carbon chain length of lipid PC 34:2 is 34. The lipid
names must be first cleaned by \code{\link{format_lipid_name}}.
}
\author{
Chenghao Zhu
}
