% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome-summarize.R
\name{nDoubleBonds}
\alias{nDoubleBonds}
\title{number of double bonds}
\usage{
nDoubleBonds(x)
}
\arguments{
\item{x}{character. The lipid annotation name. Must be first cleaned by
\code{\link{format_lipid_name}}.}
}
\value{
integer
}
\description{
This function returns the number of double bonds of a given
lipid annotation name. For example, the number of double bonds of lipid PC 34:2
is 2. The lipid annotation names must be first cleaned by
\code{\link{format_lipid_name}}.
}
\author{
Chenghao Zhu
}
