% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome-summarize.R
\name{nFattyAcyls}
\alias{nFattyAcyls}
\title{number of fatty acyls}
\usage{
nFattyAcyls(x)
}
\arguments{
\item{x}{character. The lipid annotation name. Must be first cleaned by
\code{\link{format_lipid_name}}.}
}
\value{
integer
}
\description{
This function returns the number of fatty acyls of a given
lipid annotation name. For example, the number of fatty acyl of lipid PC 34:2
is 2. The lipid annotation names must be first cleaned by
\code{\link{format_lipid_name}}.
}
\author{
Chenghao Zhu
}
