% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSet-methods.R
\name{nfeatures,HTSet-method}
\alias{nfeatures,HTSet-method}
\title{Get number of features}
\usage{
\S4method{nfeatures}{HTSet}(x)
}
\arguments{
\item{x}{\code{\link{HTSet-class}}}
}
\value{
integer
}
\description{
Get the number of features of a \code{\link{HTSet}} object
}
\seealso{
\code{\link{HTSet-class}}
}
