% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSet-methods.R
\name{nsamples,HTSet-method}
\alias{nsamples,HTSet-method}
\title{Get number of samples}
\usage{
\S4method{nsamples}{HTSet}(x)
}
\arguments{
\item{x}{\code{\link{HTSet-class}}}
}
\value{
integer
}
\description{
Get the number of samples of a \code{\link{HTSet}} object
}
\seealso{
\code{\link{HTSet-class}}
}
