% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plot.EnrichmentFET}
\alias{plot.EnrichmentFET}
\title{barplot for fet enrichment test}
\usage{
\method{plot}{EnrichmentFET}(
  x,
  greater = 3,
  less = 3,
  text.size = 10,
  fill = "gray20",
  panel.ratio = 1.2,
  ...
)
}
\arguments{
\item{x}{EnrichmentFET. Returned by \code{\link{enrichment_test}}
with test equals to fet.}

\item{greater}{integer. Number of categories which are greater.}

\item{less}{integer. Number of categories which are less.}

\item{text.size}{integer}

\item{fill}{character. Bar fill color}

\item{panel.ratio}{numeric. The ratio of left and right panels.}

\item{...}{other params not supported}
}
\value{
a ggplot object
}
\description{
barplot for fet enrichment test

This function makes bar plot for top most positively and negatively
enriched levels.
}
\examples{
lpd = transform_by_sample(lipidome, function(x) log(x/sum(x)))
design = model.matrix(~Treatment * Timepoint + Subject, data = lpd$pdata)
fit = model_fit(lpd, design, "TreatmentMed:TimepointPre", "limma")
en = enrichment_test(lpd, fit, "class", "fet")
plot(en, each = 3)
}
\seealso{
\code{\link{enrichment_test}}
}
\author{
Chenghao Zhu
}
