% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plot.EnrichmentKST}
\alias{plot.EnrichmentKST}
\title{ecdf plot for kst enrichment test}
\usage{
\method{plot}{EnrichmentKST}(x, level, alt = c("greater", "less"), colors = c("blue", "red"), ...)
}
\arguments{
\item{x}{EnrichmentKST. Returned by \code{\link{enrichment_test}}
when test equals to kst.}

\item{level}{character. The level to plot the ecdf plot. Must be in the
group.}

\item{alt}{character. Must be either greater or less.}

\item{colors}{character}

\item{...}{other arguments not supported}
}
\value{
a ggplot2 object
}
\description{
makes a ecdf plot for enrichment test result returned by
\code{\link{enrichment_test}} with test equals to kst.
}
\examples{
lpd = transform_by_sample(lipidome, function(x) log(x/sum(x)))
design = model.matrix(~Treatment * Timepoint + Subject, data = lpd$pdata)
fit = model_fit(lpd, design, "TreatmentMed:TimepointPre", "limma")
en = enrichment_test(lpd, fit, "class", "kst")
plot(en, level = "PC", alt = "greater")

}
\seealso{
\code{\link{enrichment_test}}
}
\author{
Chenghao Zhu
}
