% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSet-methods.R
\name{sampleNames,HTSet-method}
\alias{sampleNames,HTSet-method}
\alias{sampleNames}
\alias{sampleNames<-,HTSet-method}
\alias{sampleNames<-}
\title{Get or set the sample names of an HTSet object}
\usage{
\S4method{sampleNames}{HTSet}(x)

\S4method{sampleNames}{HTSet}(x) <- value
}
\arguments{
\item{x}{\code{\link{HTSet-class}}}

\item{value}{A character vector. The length must match number of samples.}
}
\value{
A character vector.
}
\description{
Get or set the sample names of an \code{\link{HTSet}} object.
}
