% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSet-methods.R
\name{subset_features}
\alias{subset_features}
\title{Subset features}
\usage{
subset_features(object, features)
}
\arguments{
\item{object}{\code{\link{HTSet-class}}}

\item{features}{The features to subset. Can be character, logical, or integers.}
}
\value{
\code{\link{HTSet-class}}
}
\description{
Get a subset of features of an HTSet object.
}
