% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome-summarize.R
\name{summarize_EOD}
\alias{summarize_EOD}
\title{Equivalent Double Bond per 18 carbons}
\usage{
summarize_EOD(object, name, class)
}
\arguments{
\item{object}{\code{\link{HTSet-class}}}

\item{name}{character. The column name of the feature variable contains the
annotation name in fdata. The annotation name must be first cleaned by the
\code{\link{format_lipid_name}}.}

\item{class}{character. The name of the feature variable that contains the
lipid class.}
}
\value{
\code{\link{HTSet-class}}
}
\description{
Calculate the \ifelse{html}{\out{EOD<sub>18</sub>}}{\eqn{EOD_{18}}}
(equivalen double bonds per 18 carbons) of each lipid class.

\ifelse{html}{EOD<sub>18</sub>}{\eqn{EOD_{18}}} of class \emph{k} in sample
\emph{i} is calculated using the equation below:

\ifelse{html}{\out{
  <center><i>
  EOD<sub>18 j,k</sub> = Σ (conc<sub>i,j,k</sub>*ndb<sub>j,k</sub>)
  / Σ (conc<sub>i,j,k</sub>*nc<sub>j,k</sub>) * 18
  </i></center>
}}{\deqn{
  EOD_{18 i,k} = (\sum conc_{i,j,k} \times ndb_{j,k}) / (\sum conc_{i,j,k} x nc_{j,8}) \times 18
}}

Where \ifelse{html}{\out{<i>conc<sub>i,j,k</sub></i>}}{\eqn{conc_{i,j,k}}}
is the mol concentration of lipid \emph{j} that belongs to class \emph{k} in
sample \emph{i}; \ifelse{html}{\out{<i>ndb<sub>j,k</sub></i>}}{\eqn{nfa_{j,k}}}
is the number of double bonds of lipid \emph{j} that belongs to class
\emph{k}; and \ifelse{html}{\out{<i>nc<sub>j,k</sub></i>}}{\eqn{nc_{j,k}}}
is the number of carbon of lipid \emph{j} that belongs to class \emph{k}.
}
\references{
Zhu, C. et al., The HDL lipidome is widely remodeled by fast food versus
Mediterranean diet in 4 days. \emph{Metabolomics}, 2019 Aug 17;15(8):114.
doi: 10.1007/s11306-019-1579-1.
}
\author{
Chenghao Zhu
}
