% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSet-methods.R
\name{summarize_feature}
\alias{summarize_feature}
\title{Summarize by feature variable}
\usage{
summarize_feature(object, feature_var)
}
\arguments{
\item{object}{\code{\link{HTSet-class}}}

\item{feature_var}{character. Must come from the column names of the
fdata. Length equals to 1.}
}
\value{
\code{\link{HTSet-class}}
}
\description{
This function summarize the edata according to the given feature
variable.
}
\author{
Chenghao Zhu
}
