% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome-summarize.R
\name{summarize_lipid_ratios}
\alias{summarize_lipid_ratios}
\title{Calculate lipid classes ratios}
\usage{
summarize_lipid_ratios(object, name, class)
}
\arguments{
\item{object}{\code{\link{HTSet-class}}}

\item{name}{character. The column name of the feature variable contains the
annotation name in fdata. The annotation name must be first cleaned by the
\code{\link{format_lipid_name}}.}

\item{class}{character. The name of the feature variable that contains the
lipid class.}
}
\value{
\code{\link{HTSet-class}}
}
\description{
Calculate ratios between lipid classes. Those calculated lipid
ratios are often indicators for enzyme activities or have diagnostic
significance. The ratios calculated are PC/LPC, PE/LPE, CE/free cholesterol,
TG/DG, SM / Cer, CE/TG, surface lipid / core lipid, PC / surface lipid, and
SM / surface lipid.
}
\author{
Chenghao Zhu
}
