% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome-summarize.R
\name{summarize_odd_chain}
\alias{summarize_odd_chain}
\title{summarize odd chain lipids}
\usage{
summarize_odd_chain(object, name, class)
}
\arguments{
\item{object}{\code{\link{HTSet-class}}}

\item{name}{character. The column name of the feature variable contains the
annotation name in fdata. The annotation name must be first cleaned by the
\code{\link{format_lipid_name}}.}

\item{class}{character. The name of the feature variable that contains the
lipid class.}
}
\value{
\code{\link{HTSet-class}}
}
\description{
Summarize odd chain lipids abundance of each lipid class.
}
\author{
Chenghao Zhu
}
