% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSet-methods.R
\name{transform_by_feature}
\alias{transform_by_feature}
\title{Transform by feature}
\usage{
transform_by_feature(object, FUN, ...)
}
\arguments{
\item{object}{\code{\link{HTSet-class}}}

\item{FUN}{function to apply.}

\item{...}{additional arguments to pass to the function}
}
\value{
\code{\link{HTSet-class}}
}
\description{
Transform the edata in a feature-wise manner (MARGIN = 2).
}
