% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{validateArgsInOptions}
\alias{validateArgsInOptions}
\title{validate args in options}
\usage{
validateArgsInOptions(args, opts)
}
\arguments{
\item{args}{list}

\item{opts}{character}
}
\description{
validate args for \code{\link{model_fit}} in given options
}
\keyword{internal}
