% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{volcanoplot.ModelFit}
\alias{volcanoplot.ModelFit}
\alias{volcanoplot}
\title{volcano plot}
\usage{
\method{volcanoplot}{ModelFit}(
  x,
  hline,
  color = "grey10",
  alpha = 0.75,
  hline_type = "dashed",
  hline_color = "grey30",
  hline_size = 0.5,
  hline_alpha = 0.75,
  ...
)
}
\arguments{
\item{x}{ModelFit. Must be returned by \code{\link{model_fit}}}

\item{hline}{numeric. The p-value where the horizontal lines to draw. All
values must between 0 and 1. If not given, serious lines will be drown at
0.05, 0.01, 0.001, 0.0001 ...}

\item{color}{character for point color}

\item{alpha}{numeric for point alpha}

\item{hline_type}{character for hline type. Default is dashed}

\item{hline_color}{chracter for hline color.}

\item{hline_size}{numeric for hline size}

\item{hline_alpha}{numeric for hline alpha}

\item{...}{other args}
}
\value{
ggplot
}
\description{
Creates a volcano plot for the returned result of
\code{\link{model_fit}}
}
\author{
Chenghao Zhu
}
