% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome-summarize.R
\name{wcmc_adduct}
\alias{wcmc_adduct}
\title{show wcmc's adduct table}
\usage{
wcmc_adduct()
}
\description{
Show West Coast Metabolomics Center's ion adduct table.
}
\seealso{
\code{\link{mz2molwt}}
}
\author{
Chenghao Zhu
}
