% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSet-methods.R
\name{subset_samples}
\alias{subset_samples}
\title{Subset samples}
\usage{
subset_samples(object, samples)
}
\arguments{
\item{object}{\code{\link{HTSet-class}}}

\item{samples}{The samples to subset. Can be character, logical, or integers.}
}
\value{
\code{\link{HTSet-class}}
}
\description{
Get a subset of samples of an HTSet object.
}
